requirements.txt
dataconf
PyYAML
ttkbootstrap